Resolving looping problems
